backend/pyproject.toml
[project]
name = "easytalk-backend"
version = "0.1.0"
description = "EasyTalk - Backend for English learning app"
authors = [
    { name = "EasyTalk Team" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",  # Совместимая с Python 3.13 версия
    "pyjwt>=2.6.0,<2.9.0",
    "python-dotenv>=0.21.0",
    "pymongo>=4.3.0,<4.6.0",
    "firebase-admin>=6.1.0,<6.3.0",
    "bcrypt>=4.0.0,<4.1.0",
    "python-multipart>=0.0.5",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "httpx==0.25.0",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
